Move traverseByDelta into SessionHistory
