+++ lib/jsdom/browser/Window.js
@@ -98,7 +98,7 @@ function Window(options) {
     document: idlUtils.implForWrapper(this._document),
     url: idlUtils.implForWrapper(this._document)._URL,
     stateObject: null
-  });
+  }, this);
 
   // TODO NEWAPI can remove this
   if (options.virtualConsole) {
