+++ lib/jsdom/living/window/History-impl.js
@@ -2,7 +2,6 @@
 const DOMException = require("../../web-idl/DOMException.js");
 const documentBaseURLSerialized = require("../helpers/document-base-url.js").documentBaseURLSerialized;
 const parseURLToResultingURLRecord = require("../helpers/document-base-url.js").parseURLToResultingURLRecord;
-const traverseHistory = require("./navigation.js").traverseHistory;
 
 exports.implementation = class HistoryImpl {
   constructor(args, privateData) {
@@ -11,8 +10,6 @@ exports.implementation = class HistoryImpl {
     this._actAsIfLocationReloadCalled = privateData.actAsIfLocationReloadCalled;
     this._state = null;
     this._latestEntry = null;
-
-    this._historyTraversalQueue = new Set();
   }
 
   _guardAgainstInactiveDocuments() {
@@ -38,22 +35,12 @@ exports.implementation = class HistoryImpl {
     this._guardAgainstInactiveDocuments();
 
     if (delta === 0) {
+      // When the go(delta) method is invoked, if delta is zero, the user agent must act as
+      // if the location.reload() method was called instead.
       this._actAsIfLocationReloadCalled();
     } else {
-      this._queueHistoryTraversalTask(() => {
-        const newIndex = this._window._sessionHistory.currentIndex + delta;
-        if (newIndex < 0 || newIndex >= this._window._sessionHistory.length) {
-          return;
-        }
-
-        const specifiedEntry = this._window._sessionHistory.entries[newIndex];
-
-        // Not implemented: unload a document guard
-
-        // Not clear that this should be queued. html/browsers/history/the-history-interface/004.html can be fixed
-        // by removing the queue, but doing so breaks some tests in history.js that also pass in browsers.
-        this._queueHistoryTraversalTask(() => traverseHistory(this._window, specifiedEntry));
-      });
+      // Otherwise, the user agent must traverse the history by a delta whose value is delta
+      this._window._sessionHistory.traverseByDelta(delta);
     }
   }
 
@@ -106,7 +93,7 @@ exports.implementation = class HistoryImpl {
     if (methodName === "pushState") {
       this._window._sessionHistory.removeAllEntriesAfterCurrentEntry();
 
-      this._clearHistoryTraversalTasks();
+      this._window._sessionHistory.clearHistoryTraversalTasks();
 
       this._window._sessionHistory.entries.push({
         document: this._document,
@@ -122,24 +109,12 @@ exports.implementation = class HistoryImpl {
       currentEntry.url = newURL;
     }
 
+    // TODO: If the current entry in the session history represents a non-GET request
+    // (e.g. it was the result of a POST submission) then update it to instead represent
+    // a GET request.
+
     this._document._URL = newURL;
     this._state = data; // TODO clone again!! O_o
     this._latestEntry = this._window._sessionHistory.currentEntry;
-  }
-
-  _queueHistoryTraversalTask(fn) {
-    const timeoutId = this._window.setTimeout(() => {
-      this._historyTraversalQueue.delete(timeoutId);
-      fn();
-    }, 0);
-
-    this._historyTraversalQueue.add(timeoutId);
-  }
-
-  _clearHistoryTraversalTasks() {
-    for (const timeoutId of this._historyTraversalQueue) {
-      this._window.clearTimeout(timeoutId);
-    }
-    this._historyTraversalQueue.clear();
   }
 };
