+++ lib/jsdom/living/window/navigation.js
@@ -116,7 +116,7 @@ exports.navigate = (window, newURL, flags) => {
 function navigateToFragment(window, newURL, flags) {
   const document = idlUtils.implForWrapper(window._document);
 
-  document._history._clearHistoryTraversalTasks();
+  window._sessionHistory.clearHistoryTraversalTasks();
 
   if (!flags.replace) {
     window._sessionHistory.removeAllEntriesAfterCurrentEntry();
